# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "qianji",
    "description": "Universal Browser Automation for AI",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/navigate": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Navigate",
        "description": "导航到 URL - 如果浏览器未运行会自动启动",
        "operationId": "navigate_api_v1_navigate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NavigateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NavigateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/snapshot": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Create Snapshot",
        "description": "创建页面快照 - 如果浏览器未运行会自动启动",
        "operationId": "create_snapshot_api_v1_snapshot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/navigate_and_snapshot": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Navigate And Snapshot",
        "description": "复合操作: 导航到URL并立即返回快照（减少 LLM 调用步骤）",
        "operationId": "navigate_and_snapshot_api_v1_navigate_and_snapshot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NavigateAndSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screenshot": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Take Screenshot",
        "description": "截图 - 如果浏览器未运行会自动启动",
        "operationId": "take_screenshot_api_v1_screenshot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pdf": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Generate Pdf",
        "description": "生成 PDF - 如果浏览器未运行会自动启动",
        "operationId": "generate_pdf_api_v1_pdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PDFRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PDFResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/go-back": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Go Back",
        "description": "后退 - 如果浏览器未运行会自动启动",
        "operationId": "go_back_api_v1_go_back_post",
        "parameters": [
          {
            "name": "browserId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Browserid"
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Targetid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/go-forward": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Go Forward",
        "description": "前进 - 如果浏览器未运行会自动启动",
        "operationId": "go_forward_api_v1_go_forward_post",
        "parameters": [
          {
            "name": "browserId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Browserid"
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Targetid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reload": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Reload",
        "description": "刷新 - 如果浏览器未运行会自动启动",
        "operationId": "reload_api_v1_reload_post",
        "parameters": [
          {
            "name": "browserId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Browserid"
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Targetid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scroll": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Scroll Page",
        "description": "滚动页面 - 支持方向滚动和滚动到指定元素",
        "operationId": "scroll_page_api_v1_scroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/get_text": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Get Element Text",
        "description": "获取指定元素的文本内容",
        "operationId": "get_element_text_api_v1_get_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/get_attribute": {
      "post": {
        "tags": [
          "snapshot"
        ],
        "summary": "Get Element Attribute",
        "description": "获取指定元素的属性",
        "operationId": "get_element_attribute_api_v1_get_attribute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAttributeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NavigateRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "browserId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Browserid"
          },
          "targetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetid"
          },
          "timeout": {
            "type": "integer",
            "title": "Timeout",
            "default": 30000
          },
          "waitUntil": {
            "type": "string",
            "title": "Waituntil",
            "default": "domcontentloaded"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "NavigateRequest"
      },
      "NavigateResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "targetId": {
            "type": "string",
            "title": "Targetid"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "snapshot": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot"
          }
        },
        "type": "object",
        "required": [
          "targetId",
          "url"
        ],
        "title": "NavigateResponse",
        "description": "导航响应"
      },
...
      "ScreenshotResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "targetId": {
            "type": "string",
            "title": "Targetid"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "path",
          "targetId",
          "url"
        ],
        "title": "ScreenshotResponse",
        "description": "截图响应"
      },
...
      "SnapshotRequest": {
        "properties": {
          "browserId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Browserid"
          },
          "targetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetid"
          },
          "refs": {
            "type": "string",
            "title": "Refs",
            "default": "role"
          },
          "maxElements": {
            "type": "integer",
            "title": "Maxelements",
            "default": 150
          },
          "viewportOnly": {
            "type": "boolean",
            "title": "Viewportonly",
            "default": false
          }
        },
        "type": "object",
        "title": "SnapshotRequest"
      },
      "SnapshotResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "targetId": {
            "type": "string",
            "title": "Targetid"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "interactive": {
            "type": "string",
            "title": "Interactive"
          },
          "domHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domhash"
          },
          "totalElements": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalelements"
          },
          "elements": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SnapshotElementData"
            },
            "type": "object",
            "title": "Elements"
          },
          "viewport": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Viewport"
          },
          "scroll": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Scroll"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "targetId",
          "url",
          "title",
          "text",
          "summary",
          "interactive",
          "elements",
          "viewport",
          "scroll"
        ],
        "title": "SnapshotResponse",
        "description": "快照响应"
      },
...
    }
  }
}